Resolviendo bugs asignacion perfil a usuario por cambio esquema de BD

+++ trunk/src/trabajo-social-web/src/main/java/gt/edu/usac/cats/controlador/usuario/ControladorAsignarPerfilUsuario.java
@@ -52,7 +52,7 @@ public class ControladorAsignarPerfilUsuario {
      * @return
      */
     @RequestMapping(value="asignarPerfilUsuario.htm", method = RequestMethod.GET)
-    public String crearFormulario(Model modelo, Short idUsuario, HttpServletRequest request) {
+    public String crearFormulario(Model modelo, Integer idUsuario, HttpServletRequest request) {
 
         if(idUsuario == null){
             return "redirect:buscarUsuario.htm";
